adds application_security_group support
this will create an ASG and associate with all VM NICs. the ID of that
ASG has been added to the module outputs so that it can be used to
enforce network routing rules (this is better that directly attaching a
nsecgrp directly to the nic). also updated ranchhand-related resources.