First working copy (#1)
* adds current work so far

this is almost complete. need to add support for security group(s) and
finish making changes to provisiong script.

* adds application_security_group support

this will create an ASG and associate with all VM NICs. the ID of that
ASG has been added to the module outputs so that it can be used to
enforce network routing rules (this is better that directly attaching a
nsecgrp directly to the nic). also updated ranchhand-related resources.

* Compatibility fixes after deployer testing (#2)

* Fix count issue, ensure ssh ready for rancher

* rke int address + ssh proxy + working_dir

* Support internal_address in rke (needed if using public ips with
  security groups)
* Add support for provisioning via bastion w/ssh proxy
* Configureable working directory to run ranchand from

* Execute ranchand remotely when using bastion

* Support optional custom data, and use CI rhel

* makes custom_data field optional

because, generally speaking, it's not required

* adds/removes comments

removing commented-out code from ranchhand launcher template
adding note about computed value change to asg association resource

* indicate ssh is ready via remote-exec echo

says it all in the code and in the terminal during "apply"

* updates ranchhand assets

segregates ranchhand variables and prefixes working_dir
joins public/private ips to leverage new ranchhand feature
bumps ranchhand version
adds a little doc to the template for posterity

* fixes remote-exec ssh user

need to use admin_username when ssh_proxy_user is not provided,
otherwise the connection will default to using "root" which does not
work.

* Connection user should always be admin username

I just flat out screwed this up initially

* bumps ranchhand version

really need to support a "latest" value so we don't have to do this
every time

* Work around tf restrictions for node ip parsing

* Make ssh_key_path bash var as well as template

We need to modify the variable when taking the bastion path in the
conditional, so it needs to be a bash variable.

* Support image version field, switch rhel ver

Was using latest, which is actually RHEL 7.6 which isn't a version
that's validated by RanchHand. I INSIST ON cAMELcASE!!!

* fixes indent

* creates load balancer to proxy traffic

variables changed/created to manage public lb/instances
moves from a list of subnet ids to a single id; fits azure networking
updated ranchhand version
adds external/internal lb (configurable via variable)
outputs ips from lb instead of instances
sets duplicated values into local variables

* changes some vm-related resource names

"this" lost some context when we added a load balancer so this change
just renames a couple of resources to make it obvious what their
intended purpose is

* bumps ranchhand version

* add output for cluster provisioning so we can `depends_on` it (#3)

* DOM-12738: Private LB Nat Support (#4)

* Add ranchhand cert name/ip support

* Split out private and public lbs

* Add completely bonkers dummy rule

* Do it properly rather than by going bonkers

BONKERS GOES NUTS: A MONKEY IN ALMONDTOWN

* fixes indentation

you know, style-guide 'n' all

* prefixes extra ranchhand variables

conforms to the naming scheme of the rest of the variables. also, move
the a ranchhand local into its respective file

* Document the comma-delimittedness of two vars